% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHME.R
\name{SHME_axes_related_metrics}
\alias{SHME_axes_related_metrics}
\title{Extract axes-related metrics of snowpatch hollows #}
\usage{
SHME_axes_related_metrics(dem, snowpatch_hollows, step = 5)
}
\arguments{
\item{dem}{object of class "SpatRast" (see \link[terra:SpatRaster-class]{terra::SpatRaster}) which is nothing else than the digital elevation model (DEM) of the area/region of interest.}

\item{snowpatch_hollows}{object of class "SpatVect" (see \link[terra:SpatVector-class]{terra::SpatVector}) with the outline(s) of the snowpatch hollow(s)}

\item{step}{Numeric. The distance between the interval nodes along the outline. Default to 5 m.}
}
\value{
snowpatch_hollow_metrics_2
A dataframe with the results of axes-related metrics extraction.
}
\description{
Extract axes-related metrics of snowpatch hollows #
}
\details{
This function allows for the calculation of the five axes-related metrics.
These are length (snowpatch hollow length (m)), width (snowpatch hollow width (m)), height (snowpatch hollow height (m)), L_W (length/width ratio (unitless)), SHS (Snowpatch hollow size, defined as the cube root of the product of length, width, and height (m)))
This function first performs the same quality checks as the function for basin-related metrics but additionally fixes topology issues, if applicable.
Obtaining the axes-related metrics requires the identification of the snowpatch hollow focus, i.e., the snowpatch hollows thresholds’ midpoints.
Note that the snowpatch hollow threshold is the relative flat part of a snowpatch hollow with slight topographical variations, corresponding to the “entrance” to a snowpatch hollow.
THis function first converts the polygon of the snowpatch hollow basin into a polyline, on which it sets interval nodes (spacing: 5 m).
To derive snowpatch hollow threshold points, the function first removes all interval nodes that have an elevation smaller than the average elevation at all interval nodes and then creates a histogram of the elevations at these points.
From this histogram, the function selects the threshold points, i.e., all interval nodes which have elevations in the most frequent elevation belt.
The next step is to select the focus point, which represents the centre of the selected threshold points.

The main axis of the snowpatch hollow should meet the condition that it divides the hollow into two halves of equal size, starting from the focus point.
The function therefore searches for the point on its backwall (target elevation: ≥ mean elevation at all points) which best satisfies this condition and creates an axis from the focus to this point.
After the determination of the hollow’s length, the function determines the width axis which passes through the midpoint of the length axis.
The function copies and rotates the length axis 90° around its midpoint, extends it, determines the intersection points with the polygon of the snow patch hollows, and calculates the distance between these two intersection points.
The determination of the hollows’ length and width allows for calculating the size of snowpatch hollows (SHS), given in m (cube root of length plus width plus height).
The height corresponds to the difference in elevation between the snowpatch hollow focus and the endpoint of the main axis on the hollow’s outline.
Finally, the function exports the four axes-related metrics as a spreadsheet creates ESRI shapefiles of the focus points, threshold points, and the axes, respectively.
}
\examples{
SHME_axes_related_metrics(dem = dem, snowpatch_hollows = snowpatch_hollows, step = 5)
}
\author{
Dr. Felix Martin Hofmann, University of Freiburg, Germany (\email{fmhofmann9892@gmail.com})
}
