% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHME.R
\name{SHME_axes_related_metrics_2}
\alias{SHME_axes_related_metrics_2}
\title{Extract axes-related metrics of snowpatch hollows with user-defined focus points}
\usage{
SHME_axes_related_metrics_2(dem, snowpatch_hollows, focus_points, step = 5)
}
\arguments{
\item{dem}{object of class "SpatRast" (see \link[terra:SpatRaster-class]{terra::SpatRaster}) which is nothing else than the digital elevation model (DEM) of the area/region of interest.}

\item{snowpatch_hollows}{object of class "SpatVect" (see \link[terra:SpatVector-class]{terra::SpatVector}) with the outline(s) of the snowpatch hollow(s)}

\item{focus_points}{object of class "SpatVect" (see \link[terra:SpatVector-class]{terra::SpatVector}) with the user-defined focus point(s) of the snowpatch hollow(s)}

\item{step}{Numeric. The distance between the interval nodes along the outline. Default to 5 m.}
}
\value{
snowpatch_hollow_metrics_2
A dataframe with the results of axes-related metrics extraction.
}
\description{
Extract axes-related metrics of snowpatch hollows with user-defined focus points
}
\details{
This function allows for the calculation of the five axes-related metrics.
These are length (snowpatch hollow length (m)), width (snowpatch hollow width (m)), height (snowpatch hollow height (m)), L_W (length/width ratio (unitless)), SHS (Snowpatch hollow size, defined as the cube root of the product of length, width, and height (m)))
The function requires the snowpatch hollow focus, i.e., the snowpatch hollows thresholds’ midpoints, as input.
Note that the snowpatch hollow threshold is the relative flat part of a snowpatch hollow with slight topographical variations, corresponding to the “entrance” to a snowpatch hollow.

The main axis of the snowpatch hollow should meet the condition that it divides the hollow into two halves of equal size, starting from the focus point.
The function therefore searches for the point on its backwall (target elevation: ≥ mean elevation at all points) which best satisfies this condition and creates an axis from the focus to this point.
After the determination of the hollow’s length, the function determines the width axis which passes through the midpoint of the length axis.
The function copies and rotates the length axis 90° around its midpoint, extends it, determines the intersection points with the polygon of the snow patch hollows, and calculates the distance between these two intersection points.
The determination of the hollows’ length and width allows for calculating the size of snowpatch hollows (SHS), given in m (cube root of length plus width plus height).
The height corresponds to the difference in elevation between the snowpatch hollow focus and the endpoint of the main axis on the hollow’s outline.
Finally, the function exports the four axes-related metrics as a spreadsheet creates ESRI shapefiles of the focus points, threshold points, and the axes, respectively.
}
\examples{
SHME_axes_related_metrics(dem = dem, snowpatch_hollows = snowpatch_hollows, focus_points = focus_points)
}
\author{
Dr. Felix Martin Hofmann, University of Freiburg, Germany (\email{fmhofmann9892@gmail.com})
}
