% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHME.R
\name{SHME_basin_related_metrics}
\alias{SHME_basin_related_metrics}
\title{Extract basin-related metrics of snowpatch hollows #}
\usage{
SHME_basin_related_metrics(dem, snowpatch_hollows)
}
\arguments{
\item{dem}{object of class "SpatRast" (see \link[terra:SpatRaster-class]{terra::SpatRaster}) which is nothing else than the digital elevation model (DEM) of the area/region of interest.}

\item{snowpatch_hollows}{object of class "SpatVect" (see \link[terra:SpatVector-class]{terra::SpatVector}) with the outline(s) of the snowpatch hollow(s)}
}
\value{
snowpatch_hollow_metrics_1
A dataframe with the exported, basin-related snowpatch hollow metrics
}
\description{
Extract basin-related metrics of snowpatch hollows #
}
\details{
This function allows for extracting 14 basin-related metrics of snowpatch hollows:
\itemize{
\item Easting
\item Northing
\item A2D (the hollows' map area)
\item elevation_min (minimum elevation in the snowpatch hollow area (m a.s.l.)
\item elevation_mean (maximum elevation in the snowpatch hollow area (m a.s.l.))
\item elevation_max (mean elevation in the snowpatch hollow area (m a.s.l.))
\item elevation_median (median elevation in the snowpatch hollow area (m a.s.l.))
\item elevation_range (Elevation range in the snowpatch hollow area (m))
\item aspect_mean (Mean aspect in the snowpatch hollow area (° from North))
\item aspect_median (Median aspect in the snowpatch hollow area (° from North))
\item slope_min (Minimum surface slope in the snowpatch hollow area (°))
\item slope_mean (Mean surface slope in the snowpatch hollow area (°))
\item slope_median (Median surface slope in the snowpatch hollow area (°))
\item slope_max (Maximum surface slope in the snowpatch hollow area (°))
}

This tool for the extraction of the 14 basin-related metrics requires a DEM (object of type "SpatRast", see) and a shapefile of the outline of the snowpatch hollows (Object of type "SpatVect", see \link[terra:SpatVector-class]{terra::SpatVector}).
This tool first performs some quality checks, i.e., it assesses whether the input-DEM and the shapefiles of the snowpatch hollow basins have the same coordinate reference system (CRS) and, if so, evaluates if the unit of the CRS is metres.
After these quality checks, the function determines the easting and the northing of the hollows’ midpoint.
For the third step, the function crops the dem to a slightly larger extent than the snowpatch hollows, derives raster files representing the aspect and the slope, and calculates the metrics associated with aspect and slope.
The function subsequently crops the DEM to the snowpatch hollows and computes the metrics related to elevation.
It finally exports the 14 basin-related snowpatch hollow metrics as a spreadsheet ("snowpatch_hollow_metrics_1.csv") to the current working directory.
}
\examples{
SHME_basin_related_metrics(dem = dem, snowpatch_hollows = snowpatch_hollows)

}
\author{
Dr. Felix Martin Hofmann, University of Freiburg, Germany (\email{fmhofmann9892@gmail.com})
}
