% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHME.R
\name{SHME_load_geodata}
\alias{SHME_load_geodata}
\title{Load geodata for snowpatch hollow metrics extraction}
\usage{
SHME_load_geodata(focus_point = FALSE)
}
\arguments{
\item{focus_point}{Logical. Should the function upload user-defined focus points? Default to FALSE.}
}
\value{
The function adds two or three objects to the global environment depending on whether "focus_point" is set to FALSE or not: "dem", "snowpatch_hollows", and optionally "focus_points".
These are of class "SpatRast", "SpatRast", and "SpatVect", respectively.
}
\description{
This function allows for loading geodata (digital elevation model (DEM), shapefile (polygons) of the snowpatch hollows, and (optionally) user-defined snowpatch hollow focus points
}
\details{
Please note that this function only accepts geodata with a coordinate reference system whose unit is metres.
At the beginning, this function makes sure that the terra package is properly installed.
This function adds two objects of SpatVector (see \link[terra:SpatVector-class]{terra::SpatVector}) representing each the outlines of the snowpatch hollows (polygon) and the focus point (if applicable) and the DEM of the area/region of interest as object of type \link[terra:SpatRaster-class]{terra::SpatRaster}.
After interactive file selection in a separate window and the upload of the data with \link[terra:rast]{terra::rast} and \link[terra:vect]{terra::vect} for the shapefile and the raster, respectively, the function performs several quality checks.
The function first assesses with \link[terra:linearUnits]{terra::linearUnits} whether the linear unit of the coordinate reference system of the DEM is metres.
The function then evaluates with \link[terra:linearUnits]{terra::linearUnits} if the linear unit of the coordinate reference system of the DEM is also metres.
It subsequently checks with \link[terra:crs]{terra::crs} if the coordinate reference systems of the input-DEM and the input-shapefile(s) are the same.
For the next step, the function assesses whether the DEM covers the whole area/region relevant for snowpatch hollow metrics calculation.
If "focus_point = TRUE", the function allows the user to interactively select the focus points for axes-related metrics calculation.
The abovementioned quality checks are then repeated for the focus point(s).
If loading geodata was successful, the function prompts the following message: "Geodata successfully imported."
}
\examples{
SHME_load_geodata(focus_point = FALSE)
}
\author{
Dr. Felix Martin Hofmann, University of Freiburg, Germany (\email{fmhofmann9892@gmail.com})
}
